Refactor to use CRISPRessoReports module
