+++ CRISPResso2/CRISPRessoCompareCORE.py
@@ -11,7 +11,7 @@ import traceback
 import argparse
 from CRISPResso2 import CRISPRessoShared
 from CRISPResso2 import CRISPRessoPlot
-from CRISPResso2 import CRISPRessoReport
+from CRISPResso2.CRISPRessoReports import CRISPRessoReport
 
 import logging
