+++ CRISPResso2/CRISPRessoMetaCORE.py
@@ -15,7 +15,7 @@ import json
 from CRISPResso2 import CRISPRessoShared
 from CRISPResso2 import CRISPRessoPlot
 from CRISPResso2 import CRISPRessoMultiProcessing
-from CRISPResso2 import CRISPRessoReport
+from CRISPResso2.CRISPRessoReports import CRISPRessoReport
 
 
 import logging
