+++ CRISPResso2/CRISPRessoPooledCORE.py
@@ -18,7 +18,7 @@ import re
 import zipfile
 from CRISPResso2 import CRISPRessoShared
 from CRISPResso2 import CRISPRessoMultiProcessing
-from CRISPResso2 import CRISPRessoReport
+from CRISPResso2.CRISPRessoReports import CRISPRessoReport
 from CRISPResso2 import CRISPRessoPlot
 import traceback
