+++ CRISPResso2/CRISPRessoPooledWGSCompareCORE.py
@@ -11,7 +11,7 @@ import os
 import sys
 from CRISPResso2 import CRISPRessoShared
 from CRISPResso2 import CRISPRessoMultiProcessing
-from CRISPResso2 import CRISPRessoReport
+from CRISPResso2.CRISPRessoReports import CRISPRessoReport
 import traceback
