+++ CRISPResso2/CRISPRessoWGSCORE.py
@@ -18,7 +18,7 @@ import traceback
 import unicodedata
 from CRISPResso2 import CRISPRessoShared
 from CRISPResso2 import CRISPRessoMultiProcessing
-from CRISPResso2 import CRISPRessoReport
+from CRISPResso2.CRISPRessoReports import CRISPRessoReport
 from CRISPResso2 import CRISPRessoPlot
